Adapt checkout shipping and address steps to new version
