+++ purchase/app/purchase/CheckoutShippingController.java
@@ -50,7 +50,7 @@ public class CheckoutShippingController extends CartController {
         final F.Promise<Cart> cartPromise = getOrCreateCart(userContext, session());
         return cartPromise.map(cart -> {
             final Messages messages = messages(userContext);
-            final CheckoutShippingPageContent content = new CheckoutShippingPageContent(cart, messages, configuration(), userContext, shippingMethods, productDataConfig, reverseRouter());
+            final CheckoutShippingPageContent content = new CheckoutShippingPageContent(cart, messages, userContext, shippingMethods, productDataConfig, reverseRouter());
             final SunrisePageData pageData = pageData(userContext, content, ctx());
             return ok(templateService().renderToHtml("checkout-shipping", pageData, userContext.locales()));
         });
@@ -59,12 +59,13 @@ public class CheckoutShippingController extends CartController {
     @AddCSRFToken
     @RequireCSRFCheck
     public F.Promise<Result> process(final String languageTag) {
+        System.out.println("A");
         final UserContext userContext = userContext(languageTag);
         return getOrCreateCart(userContext, session()).flatMap(cart -> {
             final CheckoutShippingFormData checkoutShippingFormData = extractBean(request(), CheckoutShippingFormData.class);
-            final Form<CheckoutShippingFormData> filledForm = obtainFilledForm(checkoutShippingFormData);
+            final Form<CheckoutShippingFormData> filledForm = obtainFilledForm();
             final Messages messages = messages(userContext);
-            final CheckoutShippingPageContent content = new CheckoutShippingPageContent(checkoutShippingFormData, cart, messages, configuration(), userContext, shippingMethods, productDataConfig, reverseRouter());
+            final CheckoutShippingPageContent content = new CheckoutShippingPageContent(checkoutShippingFormData, cart, messages, userContext, shippingMethods, productDataConfig, reverseRouter());
             if (filledForm.hasErrors()) {
                 return F.Promise.pure(badRequest(userContext, filledForm, content));
             } else {
@@ -75,17 +76,9 @@ public class CheckoutShippingController extends CartController {
     }
 
     private F.Promise<Cart> updateCart(final PlayJavaSphereClient sphere, final Cart cart, final CheckoutShippingFormData checkoutShippingFormData, final CheckoutShippingPageContent content) {
-        final Address shippingAddress = content.getShippingForm().getShippingAddress().toAddress();
-        final Address nullableBillingAddress = content.getShippingForm().isBillingAddressDifferentToBillingAddress()
-                ? content.getShippingForm().getBillingAddress().toAddress()
-                : null;
         final String shippingMethodId = checkoutShippingFormData.getShippingMethodId();
-        final List<UpdateAction<Cart>> updateActions = asList(
-                SetShippingAddress.of(shippingAddress),
-                SetBillingAddress.of(nullableBillingAddress),
-                SetShippingMethod.of(Reference.of(ShippingMethod.referenceTypeId(), shippingMethodId))
-        );
-        return sphere.execute(CartUpdateCommand.of(cart, updateActions));
+        final SetShippingMethod setShippingMethod = SetShippingMethod.of(Reference.of(ShippingMethod.referenceTypeId(), shippingMethodId));
+        return sphere.execute(CartUpdateCommand.of(cart, setShippingMethod));
     }
 
     private Result badRequest(final UserContext userContext, final Form<CheckoutShippingFormData> filledForm, final CheckoutShippingPageContent content) {
@@ -95,16 +88,8 @@ public class CheckoutShippingController extends CartController {
         return badRequest(templateService().renderToHtml("checkout-shipping", pageData, userContext.locales()));
     }
 
-    private Form<CheckoutShippingFormData> obtainFilledForm(final CheckoutShippingFormData checkoutShippingFormData) {
-        final Form<CheckoutShippingFormData> filledForm = Form.form(CheckoutShippingFormData.class, CheckoutShippingFormData.Validation.class).bindFromRequest(request());
-        additionalValidations(filledForm, checkoutShippingFormData);
-        return filledForm;
-    }
-
-    private void additionalValidations(final Form<CheckoutShippingFormData> filledForm, final CheckoutShippingFormData checkoutShippingFormData) {
-        if (checkoutShippingFormData.isBillingAddressDifferentToBillingAddress() && isBlank(checkoutShippingFormData.getCountryBilling())) {
-            filledForm.reject("countryBilling", "form.required");
-        }
+    private Form<CheckoutShippingFormData> obtainFilledForm() {
+        return Form.form(CheckoutShippingFormData.class, CheckoutShippingFormData.Validation.class).bindFromRequest(request());
     }
 
     private static <T> T extractBean(final Http.Request request, final Class<T> clazz) {
