+++ purchase/app/purchase/CheckoutShippingFormBean.java
@@ -1,19 +1,9 @@
 package purchase;
 
-import common.contexts.UserContext;
-import common.controllers.ReverseRouter;
 import io.sphere.sdk.carts.Cart;
-import io.sphere.sdk.models.Address;
 import io.sphere.sdk.models.Base;
-import play.Configuration;
-import play.i18n.Messages;
-
-import java.util.Optional;
 
 public class CheckoutShippingFormBean extends Base {
-    private boolean billingAddressDifferentToBillingAddress;
-    private AddressFormBean shippingAddress;
-    private AddressFormBean billingAddress;
     private ShippingMethodsFormBean shippingMethods;
     private ErrorsBean errors;
 
@@ -21,85 +11,12 @@ public class CheckoutShippingFormBean extends Base {
     public CheckoutShippingFormBean() {
     }
 
-    public CheckoutShippingFormBean(final Cart cart, final UserContext userContext, final ShippingMethods shippingMethods, final Messages messages, final Configuration configuration) {
-        final boolean billingAddressDifferentToBillingAddress = Optional.ofNullable(cart.getBillingAddress())
-                .map(Address::getLastName)
-                .map(lastName -> lastName != null)
-                .orElse(false);
-        setBillingAddressDifferentToBillingAddress(billingAddressDifferentToBillingAddress);
-        setShippingAddress(new AddressFormBean(cart.getShippingAddress(), userContext, messages, configuration));
-        setBillingAddress(new AddressFormBean(cart.getBillingAddress(), userContext, messages, configuration));
+    public CheckoutShippingFormBean(final Cart cart, final ShippingMethods shippingMethods) {
         setShippingMethods(new ShippingMethodsFormBean(cart, shippingMethods));
     }
 
-    public CheckoutShippingFormBean(final CheckoutShippingFormData checkoutShippingFormData, final UserContext userContext, final ShippingMethods shippingMethods, final Messages messages, final Configuration configuration) {
-        setBillingAddressDifferentToBillingAddress(checkoutShippingFormData.isBillingAddressDifferentToBillingAddress());
-
-        final AddressFormBean shippingAddress = createShippingAddressFormBean(checkoutShippingFormData, userContext, messages, configuration);
-        setShippingAddress(shippingAddress);
-
-        final AddressFormBean billingAddress = createBillingAddressFormBean(checkoutShippingFormData, userContext, messages, configuration);
-        setBillingAddress(billingAddress);
-
+    public CheckoutShippingFormBean(final CheckoutShippingFormData checkoutShippingFormData, final ShippingMethods shippingMethods) {
         setShippingMethods(new ShippingMethodsFormBean(checkoutShippingFormData, shippingMethods));
-    }
-
-    private AddressFormBean createShippingAddressFormBean(final CheckoutShippingFormData checkoutShippingFormData, final UserContext userContext, final Messages messages, final Configuration configuration) {
-        final AddressFormBean shippingAddress = new AddressFormBean();
-        shippingAddress.setSalutations(new SalutationsFieldsBean(checkoutShippingFormData.getTitleShipping(), messages, configuration));
-        shippingAddress.setCountries(new CountriesFieldsBean(checkoutShippingFormData.getCountryShipping(), userContext, configuration));
-        shippingAddress.setFirstName(checkoutShippingFormData.getFirstNameShipping());
-        shippingAddress.setLastName(checkoutShippingFormData.getLastNameShipping());
-        shippingAddress.setStreetName(checkoutShippingFormData.getStreetNameShipping());
-        shippingAddress.setStreetNumber(checkoutShippingFormData.getStreetNameShipping());
-        shippingAddress.setAdditionalStreetInfo(checkoutShippingFormData.getAdditionalStreetInfoShipping());
-        shippingAddress.setCity(checkoutShippingFormData.getCityShipping());
-        shippingAddress.setRegion(checkoutShippingFormData.getRegionShipping());
-        shippingAddress.setPostalCode(checkoutShippingFormData.getPostalCodeShipping());
-        shippingAddress.setPhone(checkoutShippingFormData.getPhoneShipping());
-        shippingAddress.setEmail(checkoutShippingFormData.getEmailShipping());
-        return shippingAddress;
-    }
-
-    private AddressFormBean createBillingAddressFormBean(final CheckoutShippingFormData checkoutShippingFormData, final UserContext userContext, final Messages messages, final Configuration configuration) {
-        final AddressFormBean billingAddress = new AddressFormBean();
-        billingAddress.setSalutations(new SalutationsFieldsBean(checkoutShippingFormData.getTitleBilling(), messages, configuration));
-        billingAddress.setCountries(new CountriesFieldsBean(checkoutShippingFormData.getCountryBilling(), userContext, configuration));
-        billingAddress.setFirstName(checkoutShippingFormData.getFirstNameBilling());
-        billingAddress.setLastName(checkoutShippingFormData.getLastNameBilling());
-        billingAddress.setStreetName(checkoutShippingFormData.getStreetNameBilling());
-        billingAddress.setStreetNumber(checkoutShippingFormData.getStreetNameBilling());
-        billingAddress.setAdditionalStreetInfo(checkoutShippingFormData.getAdditionalStreetInfoBilling());
-        billingAddress.setCity(checkoutShippingFormData.getCityBilling());
-        billingAddress.setRegion(checkoutShippingFormData.getRegionBilling());
-        billingAddress.setPostalCode(checkoutShippingFormData.getPostalCodeBilling());
-        billingAddress.setPhone(checkoutShippingFormData.getPhoneBilling());
-        billingAddress.setEmail(checkoutShippingFormData.getEmailBilling());
-        return billingAddress;
-    }
-
-    public AddressFormBean getBillingAddress() {
-        return billingAddress;
-    }
-
-    public void setBillingAddress(final AddressFormBean billingAddress) {
-        this.billingAddress = billingAddress;
-    }
-
-    public boolean isBillingAddressDifferentToBillingAddress() {
-        return billingAddressDifferentToBillingAddress;
-    }
-
-    public void setBillingAddressDifferentToBillingAddress(final boolean billingAddressDifferentToBillingAddress) {
-        this.billingAddressDifferentToBillingAddress = billingAddressDifferentToBillingAddress;
-    }
-
-    public AddressFormBean getShippingAddress() {
-        return shippingAddress;
-    }
-
-    public void setShippingAddress(final AddressFormBean shippingAddress) {
-        this.shippingAddress = shippingAddress;
     }
 
     public ShippingMethodsFormBean getShippingMethods() {
