+++ purchase/app/purchase/CheckoutShippingPageContent.java
@@ -4,7 +4,6 @@ import common.contexts.UserContext;
 import common.controllers.ReverseRouter;
 import common.models.ProductDataConfig;
 import io.sphere.sdk.carts.Cart;
-import play.Configuration;
 import play.i18n.Messages;
 
 public class CheckoutShippingPageContent extends CheckoutPageContent {
@@ -14,19 +13,18 @@ public class CheckoutShippingPageContent extends CheckoutPageContent {
     public CheckoutShippingPageContent() {
     }
 
-    public CheckoutShippingPageContent(final Cart cart, final Messages messages, final Configuration configuration,
-                                       final UserContext userContext, final ShippingMethods shippingMethods,
-                                       final ProductDataConfig productDataConfig, final ReverseRouter reverseRouter) {
-        fillDefaults(cart, userContext, productDataConfig, messages, reverseRouter);
-        setShippingForm(new CheckoutShippingFormBean(cart, userContext, shippingMethods, messages, configuration));
-    }
-
-    public CheckoutShippingPageContent(final CheckoutShippingFormData filledForm, final Cart cart, final Messages messages,
-                                       final Configuration configuration, final UserContext userContext,
+    public CheckoutShippingPageContent(final Cart cart, final Messages messages, final UserContext userContext,
                                        final ShippingMethods shippingMethods, final ProductDataConfig productDataConfig,
                                        final ReverseRouter reverseRouter) {
         fillDefaults(cart, userContext, productDataConfig, messages, reverseRouter);
-        setShippingForm(new CheckoutShippingFormBean(filledForm, userContext, shippingMethods, messages, configuration));
+        setShippingForm(new CheckoutShippingFormBean(cart, shippingMethods));
+    }
+
+    public CheckoutShippingPageContent(final CheckoutShippingFormData filledForm, final Cart cart, final Messages messages,
+                                       final UserContext userContext, final ShippingMethods shippingMethods,
+                                       final ProductDataConfig productDataConfig, final ReverseRouter reverseRouter) {
+        fillDefaults(cart, userContext, productDataConfig, messages, reverseRouter);
+        setShippingForm(new CheckoutShippingFormBean(filledForm, shippingMethods));
     }
 
     private void fillDefaults(final Cart cart, final UserContext userContext, final ProductDataConfig productDataConfig, final Messages messages, final ReverseRouter reverseRouter) {
