+++ purchase/app/purchase/SalutationsFieldsBean.java
@@ -3,6 +3,7 @@ package purchase;
 import common.contexts.UserContext;
 import common.models.SelectableData;
 import io.sphere.sdk.models.Address;
+import io.sphere.sdk.models.Base;
 import play.Configuration;
 import play.i18n.Messages;
 
@@ -11,7 +12,7 @@ import java.util.List;
 
 import static java.util.stream.Collectors.toList;
 
-public class SalutationsFieldsBean {
+public class SalutationsFieldsBean extends Base {
     private static final String ALLOWED_TITLES_CONFIG_KEY = "checkout.allowedTitles";
     private static final String MESSAGE_CONFIG_KEY = "message";
